+++ src/utils/strcasestr.c
@@ -28,7 +28,7 @@
 const char *
 nn_strcasestr(const char *str, const char *key)
 {
-	int len = strlen(key);
+	size_t len = strlen(key);
 
 	while (*str != '\0') {
 		if (nn_strncasecmp(str, key, len) == 0) {
